Move Queue#enqueue into separate source file
